byzantium refactored
